+++ packages/apis/manga/asura_scans/test/asura_scans_test.dart
@@ -1,6 +1,4 @@
 import 'package:asura_scans/asura_scans.dart';
 import 'package:test/test.dart';
 
-void main() async {
-
-}
+void main() async {}