+++ packages/apis/manga/cosmic_scans/test/cosmic_scans_test.dart
@@ -1,0 +1,4 @@
+import 'package:cosmic_scans/cosmic_scans.dart';
+import 'package:test/test.dart';
+
+void main() {}