+++ packages/apis/manga/void_scans/test/void_scans_test.dart
@@ -1,0 +1,4 @@
+import 'package:void_scans/void_scans.dart';
+import 'package:test/test.dart';
+
+void main() {}